refactor:Folder and case refactoring.
