+++ src/pages/login/login.tsx
@@ -3,27 +3,25 @@
 import React, { useRef } from 'react';
 
 import { ExclamationTriangleIcon, ReloadIcon } from '@radix-ui/react-icons';
-import { Button } from '../ui/button';
-import { Auth } from '../template/Auth';
+import { Button } from '../../../@/components/ui/button';
+import { Auth } from '../../templates/auth/auth';
 
-import { Alert, AlertDescription, AlertTitle } from '../ui/alert';
-import { Label } from '../ui/label';
-import { Input } from '../organism/input';
+import {
+  Alert,
+  AlertDescription,
+  AlertTitle,
+} from '../../../@/components/ui/alert';
+import { Label } from '../../../@/components/ui/label';
+import { Input } from '../../organisms/input/input';
 
-export type RegisterProps = {
+export type LoginProps = {
   company: React.ReactNode;
   onSubmitFunction: (email: string) => Promise<string>;
   variant: 'ayasofyazilim' | 'abc1';
 };
-export const Register = ({
-  variant,
-  onSubmitFunction,
-  company,
-}: RegisterProps) => {
+export const Login = ({ variant, onSubmitFunction, company }: LoginProps) => {
   const [isLoading, setIsLoading] = React.useState<boolean>(false);
   const [error, setError] = React.useState<string>('');
-  const [signInWithEmailSelected, setSignInWithEmailSelected] =
-    React.useState<boolean>(false);
 
   const emailRef = useRef<HTMLInputElement>(null);
   const passwordRef = useRef<HTMLInputElement>(null);
@@ -31,11 +29,7 @@ export const Register = ({
   async function onSubmit(event: React.SyntheticEvent) {
     event.preventDefault();
 
-    if (!emailRef.current?.value) return;
-    if (!passwordRef.current?.value) {
-      setSignInWithEmailSelected(true);
-      return;
-    }
+    if (!emailRef.current?.value || !passwordRef.current?.value) return;
     setIsLoading(true);
 
     onSubmitFunction(emailRef.current.value)
@@ -63,31 +57,18 @@ export const Register = ({
             autoFocus
           />
         </div>
-        {signInWithEmailSelected && (
-          <>
-            <div className="grid gap-1">
-              <Label htmlFor="name">Name & Surname</Label>
-              <Input
-                id="name"
-                placeholder="John Doe"
-                type="name"
-                autoComplete="name"
-                ref={emailRef}
-                autoFocus
-              />
-            </div>
-            <div className="grid gap-1">
-              <Label htmlFor="password">Password</Label>
-              <Input
-                id="password"
-                placeholder="password"
-                type="password"
-                autoComplete="no"
-                ref={passwordRef}
-              />
-            </div>
-          </>
-        )}
+
+        <div className="grid gap-1">
+          <Label htmlFor="password">Password</Label>
+          <Input
+            id="password"
+            placeholder="password"
+            type="password"
+            autoComplete="no"
+            ref={passwordRef}
+          />
+        </div>
+
         {error && (
           <Alert variant="destructive">
             <ExclamationTriangleIcon className="h-4 w-4" />
@@ -99,7 +80,7 @@ export const Register = ({
           {isLoading ? (
             <ReloadIcon className="mr-2 h-4 w-4  animate-spin" />
           ) : (
-            'Sign in with Email'
+            'Log in with Email'
           )}
         </Button>
       </div>
@@ -111,10 +92,10 @@ export const Register = ({
       variant={variant}
       formComponent={formComponent}
       isLoading={isLoading}
-      title="Create an account"
+      title="Log in"
       description={
-        <a href="/login" className="text-slate-500 text-center text-sm">
-          Do you have an account?
+        <a href="/register" className="text-slate-500 text-center text-sm">
+          Don&apos;t you have an account?
         </a>
       }
     />
