+++ src/templates/auth/auth.stories.tsx
@@ -1,7 +1,7 @@
 import React from 'react';
 import { Meta, StoryFn } from '@storybook/react';
 
-import { Auth } from '..';
+import { Auth } from '../..';
 
 export default {
   title: 'Auth Template',
